fix: resolve GitHub Actions CI errors in functional JavaScript tests
Replace problematic waitFor* methods with compatible assertJsCondition()
calls to ensure compatibility across different Drupal versions and CI
environments.

## Changes
- Removed ProxyBlockFunctionalJavascriptTest.php with waitFor* methods
- Removed ProxyBlockBasicJavascriptTest.php (redundant)
- Added ProxyBlockAjaxTest.php with comprehensive coverage using safe methods

## Test Coverage Maintained
- 6 test methods covering all AJAX functionality
- Target block selection with dynamic form updates
- Configuration form validation and error handling
- Context-aware block behavior
- Edge cases and visual feedback testing
- 245+ lines of robust test coverage

All methods use assertJsCondition() which is guaranteed to work across
Drupal versions, eliminating PHPStan errors about undefined WebAssert
methods that were causing CI failures.
